feat: remove user-side privacy-erase channel (R8)
移除用户侧「隐私紧急删除」彻底删除通道（PRD R8）：原文不留库且不对管理员
暴露，破坏举报取证与治理审计的证据留存（#549）。

- 后端：删除 POST /api/forum/user/content-privacy-erase 路由、
  PrivacyErase 控制器、PrivacyEraseContent 服务与 MessageContentPrivacyErased；
- 前端：移除 PostStream.vue / SettingsPage.vue 两处隐私紧急删除入口、
  api.ts privacyEraseContent 与四语言 i18n 键；
- 契约：openapi.yaml 删除 privacyEraseContent 操作与 PrivacyEraseRequest，
  重新生成 routes-snapshot / coverage-matrix / TS 类型，删除对应 fixture；
- 移动端：content_repository purge 移除 privacy 参数与入口，同步 l10n；
- 保留「删除 → 30 天恢复窗口 → 永久删除」主链路与 issue #492 级联下架
  （cascadeHidePostlessTopic 仍记录 privacy_delete_requested 事件）；
- 测试：privacy_erase_guard_test.go 随通道移除，级联回归测试改用
  MarkUserDeleted 构造前置状态；文档同步（identity-and-access、
  contracts-and-data、AGENTS.md）。
