+++ apps/gooseforum/app/http/routes/route4api.go
@@ -328,10 +328,9 @@ func apiRoute(ginApp *gin.Engine) {
 	forumLoginApi.POST("user/content-restore", middleware.CheckWritableAccount, middleware.RateLimit(middleware.RateLimitInteract), UpButterReq(api.RestoreContent))
 	forumLoginApi.POST("user/content-purge", middleware.CheckWritableAccount, middleware.RateLimit(middleware.RateLimitInteract), UpButterReq(api.PurgeContent))
 	// 账号/内容生命周期销毁类写路由（隐私擦除、注销账号）用放行变体：pending
-	// 用户（未验证邮箱、不打算/无法验证）仍可注销自有账号或执行紧急隐私擦除
-	// 自救，控制器自带 ownership/密码/限流校验（issue #415 review P2）；
-	// 普通写仍被 CheckWritableAccount 拦截。
-	forumLoginApi.POST("user/content-privacy-erase", middleware.CheckWritableAccountAllowPendingActivation, middleware.RateLimit(middleware.RateLimitInteract), UpButterReq(api.PrivacyErase))
+	// 用户（未验证邮箱、不打算/无法验证）仍可注销自有账号自救，控制器自带
+	// ownership/密码/限流校验（issue #415 review P2）；普通写仍被
+	// CheckWritableAccount 拦截。
 	forumLoginApi.POST("user/content-event", middleware.CheckWritableAccount, middleware.RateLimit(middleware.RateLimitInteract), UpButterReq(api.ReportContentEvent))
 	forumLoginApi.POST("user/account-close", middleware.CheckWritableAccountAllowPendingActivation, middleware.RateLimit(middleware.RateLimitInteract), UpButterReq(api.AccountClose))
 	forumLoginApi.POST("posts/create", middleware.CheckWritableAccount, middleware.RateLimit(middleware.RateLimitPostCreate), UpButterReq(api.CreatePost))
