+++ apps/gooseforum/resource/src/locales/de.ts
@@ -1136,9 +1136,6 @@ export default {
     deleteTopicTitle: 'Diese Diskussion löschen?',
     deleteTopicDescription: 'Das Thema ist nicht mehr öffentlich; falls es Antworten gibt, bleiben die Antworten anderer erhalten und dein ursprünglicher Beitrag wird zu einem Platzhalter "Inhalt vom Autor gelöscht".',
     deleteNotice: 'Du kannst gelöschte Inhalte innerhalb von 30 Tagen unter "Zuletzt gelöscht" wiederherstellen; das endgültige Löschen ist nicht umkehrbar. Bereits mit Dritten geteilte Inhalte können nicht zurückgerufen werden.',
-    privacyErase: 'Datenschutzbedenken? Beantrage die sofortige Entfernung',
-    privacyEraseConfirm: 'Eine dringende Löschung aus Datenschutzgründen beantragen? Sie überspringt die 30-tägige Wiederherstellungsfrist, entfernt Inhalte sofort überall und kann nicht rückgängig gemacht werden.',
-    privacyEraseSuccess: 'Inhalte auf deinen Antrag hin aus Datenschutzgründen entfernt.',
     authorDeletedTitle: 'Inhalt vom Autor gelöscht',
     authorDeletedPlaceholder: 'Der Autor hat diesen Inhalt gelöscht, aber die anschließende Diskussion ist weiterhin verfügbar.',
     moderatorRemovedTitle: 'Inhalt von der Moderation entfernt',
@@ -1484,16 +1481,12 @@ export default {
       retentionHint: 'Der Wiederherstellungszeitraum beträgt 30 Tage; nach dem endgültigen Löschen ist der Inhalt nicht mehr wiederherstellbar.',
       restore: 'Wiederherstellen',
       purge: 'Endgültig löschen',
-      privacyErase: 'Dringende Löschung aus Datenschutzgründen',
       restoreConfirm: 'Diesen Inhalt wiederherstellen? Nach der Wiederherstellung erscheint er wieder an seiner ursprünglichen Position.',
       purgeConfirm: 'Diesen Inhalt endgültig löschen? Diese Aktion kann nicht rückgängig gemacht werden.',
-      privacyEraseConfirm: 'Eine dringende Löschung aus Datenschutzgründen anfordern? Sie überspringt den 30-Tage-Wiederherstellungszeitraum, entfernt Inhalte sofort überall und kann nicht rückgängig gemacht werden.',
       restoreSuccess: 'Inhalt wiederhergestellt.',
       purgeSuccess: 'Inhalt endgültig gelöscht.',
-      privacyEraseSuccess: 'Inhalte auf deinen Wunsch aus Datenschutzgründen entfernt.',
       restoreHint: 'Stelle dieses Element an seiner ursprünglichen Position wieder her (30 Tage nach dem Löschen verfügbar)',
       purgeHint: 'Sofort endgültig löschen; kann nicht rückgängig gemacht werden',
-      privacyEraseHint: 'Überspringt den 30-Tage-Wiederherstellungszeitraum und entfernt sofort alles aus allen Kanälen; kann nicht wiederhergestellt werden',
     },
     status: {
       profileSaved: 'Profil gespeichert',
