+++ apps/gooseforum/resource/src/locales/en.ts
@@ -1136,9 +1136,6 @@ export default {
     deleteTopicTitle: 'Delete this topic?',
     deleteTopicDescription: 'The topic will no longer be public; if it has replies, other people\'s replies stay and your original post becomes a "Content deleted by the author" placeholder.',
     deleteNotice: 'You can restore deleted content within 30 days in "Recently deleted"; permanent deletion cannot be undone. Content already shared to third parties cannot be recalled by the community.',
-    privacyErase: 'Privacy concern? Request immediate permanent removal',
-    privacyEraseConfirm: 'Request an emergency privacy erase? It skips the 30-day recovery window, removes the content everywhere immediately, and cannot be undone.',
-    privacyEraseSuccess: 'Content removed per your privacy request.',
     authorDeletedTitle: 'Content deleted by the author',
     authorDeletedPlaceholder: 'The author deleted this content, but the follow-up discussion remains.',
     moderatorRemovedTitle: 'Content removed by moderation',
@@ -1703,16 +1700,12 @@ export default {
       retentionHint: 'The recovery window is 30 days. Permanent deletion cannot be undone.',
       restore: 'Restore',
       purge: 'Delete permanently',
-      privacyErase: 'Privacy erase',
       restoreConfirm: 'Restore this content? It will appear in its original location again.',
       purgeConfirm: 'Delete this content permanently? This cannot be undone.',
-      privacyEraseConfirm: 'Request an emergency privacy erase? It skips the 30-day recovery window, removes the content everywhere immediately, and cannot be undone.',
       restoreSuccess: 'Content restored.',
       purgeSuccess: 'Content deleted permanently.',
-      privacyEraseSuccess: 'Content removed per your privacy request.',
       restoreHint: 'Restore this item to its original place (available for 30 days after deletion)',
       purgeHint: 'Permanently delete now; this cannot be undone',
-      privacyEraseHint: 'Skips the 30-day recovery window and erases it from all channels immediately; cannot be recovered',
     },
   },
   avatarCrop: {
