+++ apps/gooseforum/resource/src/locales/ja.ts
@@ -1136,9 +1136,6 @@ export default {
     deleteTopicTitle: 'このトピックを削除しますか？',
     deleteTopicDescription: '削除後は非公開になります。返信がある場合、他のユーザーの返信は残り、本文は「投稿者が削除したコンテンツ」のプレースホルダーになります。',
     deleteNotice: '削除後30日以内は「最近削除した項目」から復元できます。完全削除後は復元できません。第三者が転載した内容はコミュニティ側で回収できません。',
-    privacyErase: 'プライバシーが心配？今すぐ完全削除を申請',
-    privacyEraseConfirm: 'プライバシー緊急削除を申請しますか？30日間の復元期間をスキップし、全チャネルから即座に完全削除され、元に戻せません。',
-    privacyEraseSuccess: 'プライバシー申請に基づきコンテンツを完全削除しました。',
     authorDeletedTitle: '投稿者が削除したコンテンツ',
     authorDeletedPlaceholder: '投稿者がこの内容を削除しましたが、その後の議論は残っています。',
     moderatorRemovedTitle: 'モデレーションにより削除されたコンテンツ',
@@ -1484,16 +1481,12 @@ export default {
       retentionHint: '復元期間は 30 日です。完全削除後は復元できません。',
       restore: '復元',
       purge: '完全削除',
-      privacyErase: 'プライバシー緊急削除',
       restoreConfirm: 'このコンテンツを復元しますか？復元すると元の位置に再表示されます。',
       purgeConfirm: 'このコンテンツを完全削除しますか？この操作は取り消せません。',
-      privacyEraseConfirm: 'プライバシー緊急削除を申請しますか？30日間の復元期間をスキップし、全チャネルから即座に完全削除され、元に戻せません。',
       restoreSuccess: 'コンテンツを復元しました。',
       purgeSuccess: 'コンテンツを完全削除しました。',
-      privacyEraseSuccess: 'プライバシー申請に基づきコンテンツを完全削除しました。',
       restoreHint: 'この内容を元の位置に復元します（削除後 30 日以内に操作できます）',
       purgeHint: 'すぐに完全削除します。取り消しはできません',
-      privacyEraseHint: '30 日間の復元期間をスキップし、すべてのチャネルから即座に完全削除します。復元はできません',
     },
     status: {
       profileSaved: 'プロフィールを保存しました',
