+++ apps/gooseforum/resource/src/locales/zh.ts
@@ -1138,9 +1138,6 @@ export default {
     deleteTopicTitle: '删除这个内容？',
     deleteTopicDescription: '删除后内容将不再公开；如有回复，其他用户的回复仍保留，你的正文会以「原帖已由作者删除」占位。',
     deleteNotice: '删除后 30 天内可在「最近删除」恢复；永久删除后无法恢复。删除内容已由第三方转发分享的，社区无法召回。',
-    privacyErase: '涉及隐私？申请立即彻底删除',
-    privacyEraseConfirm: '确定申请隐私紧急删除吗？它将跳过 30 天恢复窗口，立即从全渠道彻底删除，且不可恢复。',
-    privacyEraseSuccess: '内容已按隐私请求彻底删除。',
     authorDeletedTitle: '作者已删除内容',
     authorDeletedPlaceholder: '这段内容已由作者删除，但后续讨论仍然保留。',
     moderatorRemovedTitle: '内容已被管理移除',
@@ -1705,16 +1702,12 @@ export default {
       retentionHint: '恢复窗口为 30 天；永久删除后无法找回。',
       restore: '恢复',
       purge: '永久删除',
-      privacyErase: '隐私紧急删除',
       restoreConfirm: '确定恢复这条内容吗？恢复后它会重新出现在原来的位置。',
       purgeConfirm: '确定永久删除这条内容吗？此操作无法撤销。',
-      privacyEraseConfirm: '确定申请隐私紧急删除吗？它将跳过 30 天恢复窗口，立即从全渠道彻底删除，且不可恢复。',
       restoreSuccess: '内容已恢复。',
       purgeSuccess: '内容已永久删除。',
-      privacyEraseSuccess: '内容已按隐私请求彻底删除。',
       restoreHint: '恢复这条内容到原位置（删除后 30 天内可操作）',
       purgeHint: '立即永久删除，无法撤销',
-      privacyEraseHint: '跳过 30 天恢复窗口，立即从全渠道彻底删除，不可恢复',
     },
   },
   avatarCrop: {
