+++ apps/gooseforum/resource/src/runtime/api.ts
@@ -261,21 +261,6 @@ export async function purgeDeletedContent(contentType: DeletedContentType, conte
   return readApiResponse<boolean>(response, t('api.contentPurgeFailed'))
 }
 
-/** 隐私紧急删除（PRD R8）：跳过 30 天恢复窗口，全渠道立即彻底删除。 */
-export async function privacyEraseContent(contentType: DeletedContentType, contentId: number): Promise<boolean> {
-  const response = await fetch('/api/forum/user/content-privacy-erase', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      contentType,
-      contentId,
-    }),
-  })
-  return readApiResponse<boolean>(response, t('api.contentPurgeFailed'))
-}
-
 /** 删除生命周期埋点（PRD R14）：前端点击/确认类事件上报。 */
 export async function reportContentEvent(eventType: 'content_delete_clicked' | 'content_delete_confirmed', contentType: DeletedContentType, contentId: number): Promise<boolean> {
   const response = await fetch('/api/forum/user/content-event', {
