+++ apps/mobile/packages/forum_app/lib/l10n/app_de.arb
@@ -382,7 +382,6 @@
   "contentRestore": "Wiederherstellen",
   "contentDelete": "Löschen",
   "contentPurge": "Endgültig löschen",
-  "contentPrivacyErase": "Eigene Inhalte löschen",
   "contentDeleteConfirm": "Der Inhalt wird in den Papierkorb verschoben und kann, sofern zulässig, innerhalb von 30 Tagen wiederhergestellt werden.",
   "contentPurgeConfirm": "Dies kann nicht rückgängig gemacht werden. Text und Bilder werden endgültig gelöscht.",
   "contentPassword": "Zur Bestätigung aktuelles Passwort eingeben",
