--- packages/api-contract/fixtures/content-privacy-erase-success.json
@@ -1,5 +1,0 @@
-{
-  "result": "操作成功",
-  "code": 0,
-  "messageCode": "content.privacy.erased"
-}
